backend/requirements.txt
PyYAML~=5.1.1
Flask~=2.2.5
flask-cors~=4.0.0
beautifulsoup4~=4.11.1
requests~=2.27.1
lxml~=5.2.1